Abstracted functionality in ComparisonModal component to a helper function
